OMG OMG OMG OMG OMG OMG OMG

+++ src/bresenham.c
@@ -6,7 +6,7 @@
 /*   By: cbouwen <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/13 11:22:45 by cbouwen           #+#    #+#             */
-/*   Updated: 2023/10/17 11:20:07 by cbouwen          ###   ########.fr       */
+/*   Updated: 2023/10/18 10:22:08 by cbouwen          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ void    ft_draw_line(t_coordinates pixel_from, t_coordinates pixel_to, t_mlx_dat
     while (1)
     {
         my_mlx_pixel_put(&win_data->img, bresenham.x, bresenham.y, 0x00FF0000);
-        if (bresenham.x == pixel_to.dest_x && bresenham.y == pixel_to.dest_y)
+        if (bresenham.x == (int)pixel_to.dest_x && bresenham.y == (int)pixel_to.dest_y)
             break;
         bresenham.e2 = 2 * bresenham.err;
         if (bresenham.e2 > -bresenham.dy)
